Extract shared brain cards and add URL-routed agent history detail
Move card components (ObjectCard, ListCard, EventCard, etc.) into
brain-cards.tsx so both the sidebar brain tab and the full brains page
share the same rendering. Agent IDs now link to /activity/history/:agentId
which uses URL params instead of local state for the detail view.
